test: add topic tests
